Remove old-universal miner events
