+++ src-tauri/src/events_emitter.rs
@@ -26,7 +26,6 @@ use crate::app_in_memory_config::AppInMemoryConfig;
 use crate::events::{
     AppInMemoryConfigChangedPayload, ConnectionStatusPayload, CriticalProblemPayload,
     DisabledPhasesPayload, InitWalletScanningProgressPayload,
-    UniversalMinerInitializedExchangeIdChangedPayload,
 };
 #[cfg(target_os = "windows")]
 use crate::external_dependencies::RequiredExternalDependency;
@@ -759,24 +758,6 @@ impl EventsEmitter {
             .emit(BACKEND_STATE_UPDATE, event)
         {
             error!(target: LOG_TARGET, "Failed to emit DisabledPhasesChanged event: {:?}", e);
-        }
-    }
-
-    pub async fn emit_universal_miner_initialized_exchange_id_changed(
-        universal_miner_initialized_exchange_id: String,
-    ) {
-        let _unused = FrontendReadyChannel::current().wait_for_ready().await;
-        let event = Event {
-            event_type: EventType::UniversalMinerInitializedExchangeIdChanged,
-            payload: UniversalMinerInitializedExchangeIdChangedPayload {
-                universal_miner_initialized_exchange_id,
-            },
-        };
-        if let Err(e) = Self::get_app_handle()
-            .await
-            .emit(BACKEND_STATE_UPDATE, event)
-        {
-            error!(target: LOG_TARGET, "Failed to emit UniversalMinerInitializedExchangeIdChanged event: {:?}", e);
         }
     }
 }
