+++ src-tauri/src/setup/setup_manager.rs
@@ -26,9 +26,8 @@ use super::{
     trait_setup_phase::SetupPhaseImpl, utils::phase_builder::PhaseBuilder,
 };
 use crate::app_in_memory_config::EXCHANGE_ID;
-use crate::configs::config_core::ConfigCoreContent;
 use crate::{
-    app_in_memory_config::{DynamicMemoryConfig, ExchangeMiner, DEFAULT_EXCHANGE_ID},
+    app_in_memory_config::DEFAULT_EXCHANGE_ID,
     configs::{
         config_core::ConfigCore, config_mining::ConfigMining, config_ui::ConfigUI,
         config_wallet::ConfigWallet, trait_config::ConfigImpl,
@@ -221,7 +220,6 @@ pub struct SetupManager {
     wallet_phase_status: Sender<PhaseStatus>,
     unknown_phase_status: Sender<PhaseStatus>,
     exchange_modal_status: Sender<ExchangeModalStatus>,
-    universal_modal_status: Sender<ExchangeMiner>,
     is_app_unlocked: Mutex<bool>,
     is_wallet_unlocked: Mutex<bool>,
     is_cpu_mining_unlocked: Mutex<bool>,
@@ -881,7 +879,6 @@ impl SetupManager {
     }
 
     pub async fn start_setup(&self, app_handle: AppHandle) {
-        self.await_selected_exchange_miner(app_handle.clone()).await;
         self.pre_setup(app_handle.clone()).await;
         let _unused = self.resolve_setup_features()
             .await
@@ -895,48 +892,6 @@ impl SetupManager {
         self.setup_node_phase(app_handle.clone()).await;
         self.setup_wallet_phase(app_handle.clone()).await;
         self.setup_unknown_phase(app_handle.clone()).await;
-    }
-
-    async fn await_selected_exchange_miner(&self, app_handle: AppHandle) {
-        let state = app_handle.state::<UniverseAppState>();
-        let memory_config = state.in_memory_config.read().await;
-        let universal_miner_initialized_exchange_id = ConfigCore::content()
-            .await
-            .universal_miner_initialized_exchange_id()
-            .clone();
-        if !memory_config.is_universal_miner() || universal_miner_initialized_exchange_id.is_some()
-        {
-            return;
-        }
-        drop(memory_config);
-        let _unused = self.universal_modal_status.subscribe().changed().await;
-    }
-
-    pub async fn select_exchange_miner(
-        &self,
-        selected_miner: ExchangeMiner,
-        app_handle: AppHandle,
-    ) -> Result<(), String> {
-        let state = app_handle.state::<UniverseAppState>();
-        let mut config = state.in_memory_config.write().await;
-        let new_config = DynamicMemoryConfig::init_exchange_miner(&selected_miner.id);
-        let new_config_cloned = new_config.clone();
-        *config = new_config;
-
-        EventsEmitter::emit_app_in_memory_config_changed(new_config_cloned, true).await;
-        let _unused = ConfigCore::update_field(
-            ConfigCoreContent::set_universal_miner_initialized_exchange_id,
-            Some(selected_miner.id.clone()),
-        )
-        .await;
-        EventsEmitter::emit_universal_miner_initialized_exchange_id_changed(
-            selected_miner.id.clone(),
-        )
-        .await;
-        self.universal_modal_status
-            .send(selected_miner.clone())
-            .map_err(|e| e.to_string())?;
-        Ok(())
     }
 
     pub async fn handle_switch_to_local_node(&self) {
