+++ src/hooks/app/useTauriEventsListener.ts
@@ -51,7 +51,6 @@ import {
     handleConfigMiningLoaded,
     handleConfigUILoaded,
     handleConfigWalletLoaded,
-    handleUniversalMinerInitializedExchangeIdChanged,
     handleMiningTimeUpdate,
 } from '@app/store/actions/appConfigStoreActions';
 import { invoke } from '@tauri-apps/api/core';
@@ -218,9 +217,6 @@ const useTauriEventsListener = () => {
                             break;
                         case 'DisabledPhasesChanged':
                             handleUpdateDisabledPhases(event.payload);
-                            break;
-                        case 'UniversalMinerInitializedExchangeIdChanged':
-                            handleUniversalMinerInitializedExchangeIdChanged(event.payload);
                             break;
                         default:
                             console.warn('Unknown event', JSON.stringify(event));
