+++ src/store/actions/appConfigStoreActions.ts
@@ -27,10 +27,7 @@ import { ConfigBackendInMemory, ConfigCore, ConfigMining, ConfigUI, ConfigWallet
 import { NodeType, updateNodeType as updateNodeTypeForNodeStore } from '../useNodeStore.ts';
 import { fetchExchangeContent, fetchExchangeMiners, setShowUniversalModal } from '../useExchangeStore.ts';
 
-import {
-    AppInMemoryConfigChangedPayload,
-    UniversalMinerInitializedExchangeIdChangedPayload,
-} from '@app/types/events-payloads.ts';
+import { AppInMemoryConfigChangedPayload } from '@app/types/events-payloads.ts';
 
 interface SetModeProps {
     mode: modeType;
@@ -332,7 +329,7 @@ export const fetchBackendInMemoryConfig = async () => {
             const universalMinerExchangeId = await invoke('get_universal_miner_initialized_exchange_id'); // It has to be a command instead of getter from appConfigStore since store is not initialized yet
             const isUniversalUninitialized = isUniversalMiner && !universalMinerExchangeId;
             const isUniversalInitialized = isUniversalMiner && universalMinerExchangeId;
-            const isExchangeMode = res.exchangeId && !isUniversalMiner && res.exchangeId !== 'classic';
+            const isExchangeMode = res.exchangeId && !isUniversalMiner && res.exchangeId !== 'universal';
             if (isUniversalUninitialized) {
                 await fetchExchangeMiners();
                 setShowUniversalModal(true);
@@ -349,18 +346,8 @@ export const fetchBackendInMemoryConfig = async () => {
     }
 };
 
-export const handleUniversalMinerInitializedExchangeIdChanged = (
-    payload: UniversalMinerInitializedExchangeIdChangedPayload
-) => {
-    useConfigCoreStore.setState({
-        universal_miner_initialized_exchange_id: payload.universal_miner_initialized_exchange_id,
-    });
-    if (payload.universal_miner_initialized_exchange_id) {
-        setShowUniversalModal(false); // Enforce this flag if there is race condition between this and handleAppInMemoryConfigChanged
-    }
-};
-
 export const handleAppInMemoryConfigChanged = (payload: AppInMemoryConfigChangedPayload) => {
+    console.info('[DEBUG EXCHANGE MINER] AppInMemoryConfigChangedPayload ', payload);
     const newConfig: ConfigBackendInMemory = {
         airdropApiUrl: payload.app_in_memory_config.airdrop_api_url,
         airdropUrl: payload.app_in_memory_config.airdrop_url,
