+++ src/types/configs.ts
@@ -23,7 +23,6 @@ export interface ConfigCore {
     };
     remote_base_node_address: string;
     node_type?: NodeType;
-    universal_miner_initialized_exchange_id?: string;
 }
 export interface ConfigWallet {
     created_at: string;
